+++ CHANGELOG.md
@@ -28,3 +28,4 @@ None.
 - Added an automated Docker runner for tests that can run on GPUs. See [testing/runner/README.md](testing/runner/README.md) for more information.
 - When installing package with pip, the commit hash and date of the installation is saved into `drugex._version`. This information is also used as a basis of a new dynamic versioning scheme for the package. The version number is generated automatically upon installation of the package and saved to `drugex.__version__`. 
 - QSPRPred is now available as an optional dependency that can be installed with DrugEx using the `[qsprpred]` option.
+- The `FileMonitor` now has a attribute `save_model_option` which can be "all", "improvement" and "best" instead of only having the "best" option.